HBD-16: Refactored; (c) ContactCloudFragment & (c) ContactBackupFragment; Renamed.
